+++ dev-packages/cli/src/malagu-main.ts
@@ -1,6 +1,6 @@
 import { fork, ChildProcess } from 'child_process';
 import { Component, Module } from '@malagu/cli-common';
-import { join, sep } from 'path';
+import { sep } from 'path';
 const Watchpack = require('watchpack');
 
 const watchpack = new Watchpack({});
@@ -12,12 +12,14 @@ argv.shift();
 let current: ChildProcess;
 
 interface Data {
-    components: Component[];
-    configHookModules: Module[];
-    webpackHookModules: Module[];
-    serveHookModules: Module[];
-    buildHookModules: Module[];
-    deployHookModules: Module[];
+    components?: Component[];
+    configHookModules?: Module[];
+    webpackHookModules?: Module[];
+    serveHookModules?: Module[];
+    buildHookModules?: Module[];
+    deployHookModules?: Module[];
+    propsHookModules?: Module[];
+    infoHookModules?: Module[];
 }
 
 const exitListener = (code: number | null) => {
@@ -35,20 +37,24 @@ function execute() {
         current.removeListener('exit', exitListener);
         current.kill();
     }
-    current = fork(join(__dirname, 'malagu.js'), argv, { stdio: 'inherit' });
+    const malaguPath = require.resolve('@malagu/cli/lib/malagu', { paths: [ process.cwd(), __dirname ] });
+    current = fork(malaguPath, argv, { stdio: 'inherit' });
     // eslint-disable-next-line no-null/no-null
     current.on('exit', exitListener);
     current.on('error', () => process.exit(-1));
     current.on('message', (messageEvent: MessageEvent<Data>) => {
         if (messageEvent.type === 'cliContext') {
-            const { components, webpackHookModules, configHookModules, buildHookModules, serveHookModules, deployHookModules } = messageEvent.data;
+            const { components, webpackHookModules, configHookModules, buildHookModules,
+                serveHookModules, deployHookModules, propsHookModules, infoHookModules } = messageEvent.data;
             const files = [
-                ...webpackHookModules.map(m => m.path),
-                ...configHookModules.map(m => m.path),
-                ...buildHookModules.map(m => m.path),
-                ...serveHookModules.map(m => m.path),
-                ...deployHookModules.map(m => m.path),
-                ...components.reduce<string[]>((prev, curr) => prev.concat(curr.configFiles), [])
+                ...(webpackHookModules || []).map(m => m.path),
+                ...(configHookModules || []).map(m => m.path),
+                ...(buildHookModules || []).map(m => m.path),
+                ...(serveHookModules || []).map(m => m.path),
+                ...(deployHookModules || []).map(m => m.path),
+                ...(propsHookModules || []).map(m => m.path),
+                ...(infoHookModules || []).map(m => m.path),
+                ...(components || []).reduce<string[]>((prev, curr) => prev.concat(curr.configFiles), [])
             ];
             watchpack.watch({
                 files: files.map(f => f.split('/').join(sep)),